+++ src/dotnet/ReSharperPlugin.HeapView.Tests/BoxingAllocationsTest.cs
@@ -76,6 +76,11 @@ public abstract class BoxingAllocationsTestBase : CSharpHighlightingTestBase
   [Test] public void TestExtensionMethodBoxing04() { DoNamedTest2(); }
 
   [Test] public void TestArgList01() { DoNamedTest2(); }
+
+  [Test] public void TestLinqCast01() { DoNamedTest2(); }
+  [Test] public void TestLinqCast02() { DoNamedTest2(); }
+  [Test] public void TestLinqCast03() { DoNamedTest2(); }
+  [Test] public void TestLinqCast04() { DoNamedTest2(); }
 }
 
 [TestNetFramework46]