+++ src/dotnet/ReSharperPlugin.HeapView.Tests/test/data/Boxing/LinqCast01.cs
@@ -1,0 +1,8 @@
+using System.Linq;
+
+var xs = new int[0];
+
+var ys = from object o in xs
+         from object v in xs
+         join object u in xs on o equals u
+         select o;