+++ src/dotnet/ReSharperPlugin.HeapView.Tests/test/data/Boxing/LinqCast03.cs
@@ -1,0 +1,4 @@
+using System.Linq;
+
+var xs = new int[0];
+var ys = xs.Cast<object>();