+++ src/dotnet/ReSharperPlugin.HeapView.Tests/test/data/Boxing/LinqCast04.cs
@@ -1,0 +1,6 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+var xs = new List<ConsoleKey>();
+var ys = Enumerable.Cast<Enum>(xs);